thelab/thelab_backend/requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pep8
pylint